python/requirements.txt
matplotlib==3.1.1
mip==1.5.3
networkx==2.3
numpy==1.17.2
pandas==0.25.3
plotly==4.3.0
pycparser==2.19
pyparsing==2.4.2
PyYAML==5.1.2
resource-retriever==1.12.6
retrying==1.3.3
scipy==1.3.1
seaborn==0.10.0
six==1.12.0
xmltodict==0.12.0
